backend/requirements.txt
fastapi~=0.115
uvicorn~=0.34
asyncpg~=0.30
pydantic~=2.11
boto3~=1.37
python-jose[cryptography]~=3.5
requests~=2.31
pyyaml~=6.0
pillow~=10.0
beautifulsoup4~=4.12
tqdm~=4.67
supabase~=2.15
python-dotenv~=1.0
pytest-asyncio
colorama
pydantic[email]
APScheduler
openai
pytz
python-multipart
bcrypt>=4
aiohttp
google-cloud-vision
numpy
google-auth
